Add script to get coor. and map of park locations
Wrote a script that uses the park names of the dataset in the DB to
get the latitude and longitude coordinates using Google Geocoding API.
Currently, these coordinates are just saved in a json file and the
map is importing these locations from the file. We will add to
DB in the future.

The map is currently (temporarily) at /parks/map endpoint.