Migrate from Folium to using LeafletJS directly for mapping
Folium is a python wrapper for the JS library leaflet. Folium only
implements subset of leaflet and converts the Python code to html.
As such, Folium actually has a lot less functionality and makes
manipulating elements with JS and the DOM very hard. Like connecting the
parks in the park list with individual markers on the map. I couldn't
do it because there is no way to pass each marker an ID for connection.

This commit changes from using Folium to using LeafletJS for the map
directly. Does not change any other part of the DJango system, just the
map. It is now written in JS.

This commit implements the basics of the map:
- markers displaying
- highlighting markers when hovering over park
- centering and zoom control
- Other UI/UX stuff.