Fix star icon issue
It was due to the fallback font family the browser chose. The fallback
font could not render the stars. I am not sure why the browser needed
to use a fallback font in the first place as the specified font family
should have been available.

To avoid other users potentially having the same problem, I changed the
stars from a unicode character to being an svg image.

Also added better rating displays in details page.

Added star displays on the map page in the park list and in the map
popup.